requirements.txt
python-dotenv==1.0.1
websockets==12.0
openai==1.57.2
